Adds Secrets to Repo
This gives us a simple way to pass secrets and environment files back and forth. It requires a single password. I didn't want to mess with keys and all that, primarily because I don't understand it, but also because I don't think it's necessary for our purposes.

I think this will be a major win in helping us manage all the secrets for the different churches.