moved MATCH_APP_IDENTIFIER to env
I liked the idea of it being automatic at first, but in practice it was very confusing when trying to generate new certificates because you couldn't use the magic from inside the lane and therefore needed to define it in the env file anyway. this removes complexity by being more explicit and allows us to reuse the app ID variables for other fastlane commands locally.